requirements.txt
Django==5.1
pytesseract
opencv-python
python-docx
weasyprint
requests
Pillow